backend/dogback/requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.1.1
cryptography==37.0.2
cycler==0.11.0
defusedxml==0.7.1
dj-database-url==0.5.0
dj-rest-auth==2.2.5
Django==3.2
django-admin-honeypot==1.1.0
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-haystack==3.2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.1
environ==1.0
fonttools==4.33.3
gunicorn==20.1.0
h11==0.13.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.2
mysqlclient==2.1.1
numpy==1.22.4
oauthlib==3.2.1
outcome==1.1.0
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
types-cryptography==3.3.23
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.2.0
Whoosh==2.7.4
wsproto==1.1.0
